Move test extensions
Move test extensions and other functions for ease of testing into a TestHelper file. Also, update for Create and Update flows and introduce simpler asset methods for recording (or not recording) a transaction.
